+++ src/main/java/com/nisovin/shopkeepers/SKShopkeepersPlugin.java
@@ -34,6 +34,7 @@ import com.nisovin.shopkeepers.api.shopkeeper.offers.PriceOffer;
 import com.nisovin.shopkeepers.api.shopkeeper.offers.TradingOffer;
 import com.nisovin.shopkeepers.api.shopkeeper.player.PlayerShopkeeper;
 import com.nisovin.shopkeepers.chestprotection.ProtectedChests;
+import com.nisovin.shopkeepers.chestprotection.RemoveShopOnChestBreak;
 import com.nisovin.shopkeepers.commands.Commands;
 import com.nisovin.shopkeepers.compat.NMSManager;
 import com.nisovin.shopkeepers.config.ConfigLoadException;
@@ -113,6 +114,7 @@ public class SKShopkeepersPlugin extends JavaPlugin implements ShopkeepersPlugin
 	private final ShopkeeperCreation shopkeeperCreation = new ShopkeeperCreation(this);
 
 	private final ProtectedChests protectedChests = new ProtectedChests(this);
+	private final RemoveShopOnChestBreak removeShopOnChestBreak = new RemoveShopOnChestBreak(this, protectedChests);
 	private final LivingShops livingShops = new LivingShops(this);
 	private final SignShops signShops = new SignShops(this);
 	private final CitizensShops citizensShops = new CitizensShops(this);
@@ -338,6 +340,7 @@ public class SKShopkeepersPlugin extends JavaPlugin implements ShopkeepersPlugin
 
 		// enable ProtectedChests:
 		protectedChests.enable();
+		removeShopOnChestBreak.onEnable();
 
 		// register events:
 		PluginManager pm = Bukkit.getPluginManager();
@@ -456,6 +459,7 @@ public class SKShopkeepersPlugin extends JavaPlugin implements ShopkeepersPlugin
 
 		// disable protected chests:
 		protectedChests.disable();
+		removeShopOnChestBreak.onDisable();
 
 		// disable shopkeeper registry: unloads all shopkeepers
 		shopkeeperRegistry.onDisable();
@@ -562,6 +566,12 @@ public class SKShopkeepersPlugin extends JavaPlugin implements ShopkeepersPlugin
 
 	public ProtectedChests getProtectedChests() {
 		return protectedChests;
+	}
+
+	// SHOPKEEPR REMOVAL ON CHEST BREAKING
+
+	public RemoveShopOnChestBreak getRemoveShopOnChestBreak() {
+		return removeShopOnChestBreak;
 	}
 
 	// LIVING ENTITY SHOPS